Fix the type of the forward argument in cowngc1/cowngc4.
The test_cown_gc function expects a uint64_t as its first argument, but
we were passing a size_t. Because there's an indirection through a
template, and we're letting the arguments get deduced, implicit
conversion did not kick in.

This was causing a build failure on macOS for the runtime tests. On
there uint64_t and size_t are different incompatible types, despite both
being 64-bit and unsigned integers :roll_eyes:.